refactor: Separate schemas into actionsSchema and configSchema
